requirements.txt
streamlit==1.44.0
pandas==2.2.2
seaborn==0.13.2
matplotlib
numpy
scikit-learn==1.6.1
plotly
pillow